Refactor testReturnNumberValue and requiredVTextField
The test isn't evaluated properply as v-text-field input event
will not be emitted if there is changes on bound model or on props
changes of parent component. It needs to be triggered manually.

This commit refactors testReturnNumberValue to mock changes of user
input by updating reactivity data (targetItem) then manually trigger input
event of v-text-field. After that, on-input-change event of parameter-input
will be emitted.